rename to ready for headline in detail-area
