+++ linen-glue/src/main/java/x7c1/linen/glue/res/layout/UnreadDetailRowEntry.java
@@ -1,0 +1,32 @@
+
+/**
+ * This file is automatically generated by wheat-build.
+ * Do not modify this file -- YOUR CHANGES WILL BE ERASED!
+ */
+
+package x7c1.linen.glue.res.layout;
+
+import android.view.View;
+import android.widget.TextView;
+
+public class UnreadDetailRowEntry extends UnreadDetailRow {
+
+    public final TextView title;
+    public final TextView content;
+    public final TextView createdAt;
+    public final TextView visit;
+
+    public UnreadDetailRowEntry(
+        View itemView,
+        TextView title,
+        TextView content,
+        TextView createdAt,
+        TextView visit
+    ){
+        super(itemView);
+        this.title = title;
+        this.content = content;
+        this.createdAt = createdAt;
+        this.visit = visit;
+    }
+}
