+++ linen-modern/src/main/scala/x7c1/linen/modern/init/unread/UnreadItemsDelegatee.scala
@@ -6,7 +6,7 @@ import android.util.TypedValue
 import android.util.TypedValue.COMPLEX_UNIT_DIP
 import android.view.KeyEvent
 import x7c1.linen.glue.activity.ActivityControl
-import x7c1.linen.glue.res.layout.{UnreadOutlineRowSource, UnreadDetailRow, UnreadOutlineRowEntry, MainLayout, MenuRowLabel, MenuRowTitle, MenuRowSeparator, UnreadSourceRow}
+import x7c1.linen.glue.res.layout.{UnreadOutlineRowSource, UnreadDetailRowEntry, UnreadOutlineRowEntry, MainLayout, MenuRowLabel, MenuRowTitle, MenuRowSeparator, UnreadSourceRow}
 import x7c1.linen.modern.accessor.{EntryAccessor, LinenOpenHelper, RawSourceAccessor, UnreadSourceAccessor}
 import x7c1.linen.modern.display.unread.{PaneContainer, DetailArea, OutlineArea, SourceArea}
 import x7c1.linen.modern.struct.{UnreadDetail, UnreadOutline}
@@ -115,5 +115,5 @@ class UnreadRowProviders(
   val forSource: ViewHolderProvider[UnreadSourceRow],
   val forOutlineSource: ViewHolderProvider[UnreadOutlineRowSource],
   val forOutlineEntry: ViewHolderProvider[UnreadOutlineRowEntry],
-  val forDetail: ViewHolderProvider[UnreadDetailRow]
+  val forDetail: ViewHolderProvider[UnreadDetailRowEntry]
 )
