+++ linen-starter/src/main/java/x7c1/linen/res/layout/UnreadDetailRowEntryProvider.java
@@ -1,0 +1,47 @@
+
+/**
+ * This file is automatically generated by wheat-build.
+ * Do not modify this file -- YOUR CHANGES WILL BE ERASED!
+ */
+
+package x7c1.linen.res.layout;
+
+import android.content.Context;
+import android.view.LayoutInflater;
+import android.view.ViewGroup;
+import android.view.View;
+import android.widget.TextView;
+import x7c1.wheat.ancient.resource.ViewHolderProvider;
+import x7c1.linen.R;
+import x7c1.linen.glue.res.layout.UnreadDetailRowEntry;
+
+public class UnreadDetailRowEntryProvider implements ViewHolderProvider<UnreadDetailRowEntry> {
+
+    private final LayoutInflater inflater;
+
+    public UnreadDetailRowEntryProvider(Context context){
+        inflater = LayoutInflater.from(context);
+    }
+
+    @Override
+    public int layoutId(){
+        return R.layout.unread_detail_row__entry;
+    }
+
+    @Override
+    public UnreadDetailRowEntry inflateOn(ViewGroup parent){
+        return inflate(parent, false);
+    }
+
+    @Override
+    public UnreadDetailRowEntry inflate(ViewGroup parent, boolean attachToRoot){
+        View view = inflater.inflate(R.layout.unread_detail_row__entry, parent, attachToRoot);
+        return new UnreadDetailRowEntry(
+            view,
+            (TextView) view.findViewById(R.id.unread_detail_row__entry__title),
+            (TextView) view.findViewById(R.id.unread_detail_row__entry__content),
+            (TextView) view.findViewById(R.id.unread_detail_row__entry__created_at),
+            (TextView) view.findViewById(R.id.unread_detail_row__entry__visit)
+        );
+    }
+}
