--- linen-starter/src/main/java/x7c1/linen/res/layout/UnreadDetailRowProvider.java
@@ -1,47 +1,0 @@
-
-/**
- * This file is automatically generated by wheat-build.
- * Do not modify this file -- YOUR CHANGES WILL BE ERASED!
- */
-
-package x7c1.linen.res.layout;
-
-import android.content.Context;
-import android.view.LayoutInflater;
-import android.view.ViewGroup;
-import android.view.View;
-import android.widget.TextView;
-import x7c1.wheat.ancient.resource.ViewHolderProvider;
-import x7c1.linen.R;
-import x7c1.linen.glue.res.layout.UnreadDetailRow;
-
-public class UnreadDetailRowProvider implements ViewHolderProvider<UnreadDetailRow> {
-
-    private final LayoutInflater inflater;
-
-    public UnreadDetailRowProvider(Context context){
-        inflater = LayoutInflater.from(context);
-    }
-
-    @Override
-    public int layoutId(){
-        return R.layout.unread_detail_row;
-    }
-
-    @Override
-    public UnreadDetailRow inflateOn(ViewGroup parent){
-        return inflate(parent, false);
-    }
-
-    @Override
-    public UnreadDetailRow inflate(ViewGroup parent, boolean attachToRoot){
-        View view = inflater.inflate(R.layout.unread_detail_row, parent, attachToRoot);
-        return new UnreadDetailRow(
-            view,
-            (TextView) view.findViewById(R.id.unread_detail_row__title),
-            (TextView) view.findViewById(R.id.unread_detail_row__content),
-            (TextView) view.findViewById(R.id.unread_detail_row__created_at),
-            (TextView) view.findViewById(R.id.unread_detail_row__visit)
-        );
-    }
-}
