+++ linen-starter/src/main/res/layout/unread_detail_row__entry.xml
@@ -11,7 +11,7 @@
         style="@style/AppCardViewRow"
         >
         <TextView
-            android:id="@+id/unread_detail_row__title"
+            android:id="@+id/unread_detail_row__entry__title"
             android:textSize="17sp"
             android:layout_marginTop="20dp"
             android:layout_width="wrap_content"
@@ -26,12 +26,12 @@
             android:id="@+id/border1"
             android:layout_width="match_parent"
             android:layout_height="1px"
-            android:layout_below="@id/unread_detail_row__title"
+            android:layout_below="@id/unread_detail_row__entry__title"
             android:background="@color/text_tertiary"
             />
 
         <TextView
-            android:id="@+id/unread_detail_row__content"
+            android:id="@+id/unread_detail_row__entry__content"
             android:textSize="15sp"
             android:layout_marginTop="12dp"
             android:layout_width="wrap_content"
@@ -44,13 +44,13 @@
             />
 
         <TextView
-            android:id="@+id/unread_detail_row__created_at"
+            android:id="@+id/unread_detail_row__entry__created_at"
             android:textSize="12sp"
             android:textAlignment="textEnd"
             android:layout_marginTop="12dp"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_below="@id/unread_detail_row__content"
+            android:layout_below="@id/unread_detail_row__entry__content"
             android:textColor="@color/text_tertiary"
             />
 
@@ -58,12 +58,12 @@
             android:id="@+id/border2"
             android:layout_width="match_parent"
             android:layout_height="1px"
-            android:layout_below="@id/unread_detail_row__created_at"
+            android:layout_below="@id/unread_detail_row__entry__created_at"
             android:background="@color/text_tertiary"
             />
 
         <TextView
-            android:id="@+id/unread_detail_row__visit"
+            android:id="@+id/unread_detail_row__entry__visit"
             android:layout_width="wrap_content"
             android:layout_height="45dp"
             android:layout_below="@id/border2"
